requirements.txt
ag2==0.8.5
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.5.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
duckduckgo_search==8.0.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
FLAML==2.3.4
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.10.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.2
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
openai==1.72.0
packaging==24.2
pillow==11.1.0
primp==0.14.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.8.5
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
rsa==4.9
sgmllib3k==1.0.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==3.0.1
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
vertexai==1.71.1
websockets==15.0.1
